requirements.txt
# BioFace AI - Real-Time Behavioral Intelligence System
# Dependências principais do projeto

# Visão Computacional
opencv-python==4.8.1.78  # Versão compatível com NumPy < 2.0
mediapipe==0.10.7
numpy==1.26.4  # MediaPipe requer NumPy < 2.0 (compatibilidade)
protobuf<5.0,>=4.25.3  # MediaPipe requer protobuf < 5.0
Pillow==10.1.0

# Machine Learning / Deep Learning (opcional - apenas se usar DeepFace)
# tensorflow==2.15.0  # Descomente apenas se for usar DeepFace
# keras==2.15.0
scikit-learn==1.3.2
# deepface==0.0.79  # Descomente apenas se for usar DeepFace (requer TensorFlow)

# Backend API (para fases futuras)
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6

# Banco de Dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging e Monitoramento
loguru==0.7.2

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0  # Cobertura de código
pytest-mock==3.12.0  # Mocks para testes

# Dashboard
streamlit==1.28.2
plotly==5.18.0
requests==2.31.0  # Para dashboard fazer requisições à API


